Make sure component files have pascal case
