+++ app/analytics/page.tsx
@@ -1,4 +1,4 @@
-import { AnalyticsContent } from './analytics-content'
+import { AnalyticsContent } from './AnalyticsContent'
 import { AnalyticsService } from '../../services/analytics'
 
 export default async function Analytics() {
