--- components/GridItem.tsx
@@ -1,58 +1,0 @@
-import NextLink from 'next/link'
-import Image from 'next/image'
-import { Box, Text, LinkBox, LinkOverlay, Heading } from '@chakra-ui/react'
-import { Global } from '@emotion/react'
-
-export const PostGridItem = ({ children, href, title, thumbnail }) => {
-  return (
-    <Box w="100%" alignItems="center">
-      <LinkBox cursor="pointer">
-        <Image
-          src={`http:${thumbnail.url}`}
-          alt={title}
-          layout=""
-          width={thumbnail.details.image.width}
-          height={thumbnail.details.image.height}
-          className="grid-item-thumbnail"
-          loading="lazy"
-        />
-        <LinkOverlay href={href} target="_blank">
-          <Text mt={2} fontSize={['md', 'xl', '3xl']} fontWeight="bold">
-            {title}
-          </Text>
-        </LinkOverlay>
-        <Text>{children}</Text>
-      </LinkBox>
-    </Box>
-  )
-}
-
-export const ProjectGridItem = ({ children, slug, title, thumbnail }) => (
-  <Box w="100%" alignItems="center">
-    <NextLink href={'/projects/' + slug} passHref>
-      <LinkBox cursor="pointer">
-        <Image
-          src={`http:${thumbnail.url}`}
-          alt={title}
-          width={thumbnail.details.image.width}
-          height={thumbnail.details.image.height}
-          className="grid-item-thumbnail"
-        />
-        <Heading textAlign="center" textStyle="h3" my={2} fontWeight="bold">
-          {title}
-        </Heading>
-        <Text>{children}</Text>
-      </LinkBox>
-    </NextLink>
-  </Box>
-)
-
-export const GridItemStyle = () => (
-  <Global
-    styles={`
-  .grid-item-thumbnail {
-    border-radius:12px
-  }
-  `}
-  />
-)
